requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
click==8.1.8
decouple==0.0.7
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
uvicorn==0.34.0
whitenoise==6.9.0
